Renamed io-wallet-user-func/src/infra/http/whitelisted-fiscal-code.ts
